update PR 153 with critical edits
As mentioned in various comments in PR #153:

    https://github.com/lcn2/calc/pull/153

we needed to modify the PR you sent us for various reasons,
including but not limited to having to satiny a number
of older systems on which calc v2 needs to run.

Please examine these "mods to your mods" and see if
they are OK form your perspective.   Feel free to update
your PR if you find any issues.

Once you approve or update your PR, we will complete our review.