prep for pre-v3 fork code reformat
Added a new `.clang-format` file as the basis for using the
`clang-format(1)` tool to format calc source in a consistent way.
This code re-formatting will make it easier to work in calc v3 code
changes in the future by making the calc source much more consistent.

Removed `.clang-format` exclusion from `.gitignore`.

Modified `func.c` to prevent the `clang-format(1)` tool from formatting
the `builtins[]` table that builtin make rule in `help/Makefile` uses to
format the `help/builtin` file.  We can live with the `builtins[]`
table as it is today, for now.

The initial new `.clang-format` file was successfully tested by:

```sh
    make clobber
    find . -name '*.[ch]' -print | xargs clang-format -i
    make clobber all chk
```

The changes made by the `clang-format(1)` tool have **NOT** be
checked in: we are still evaluating the impact.

There are **NO** API changes with this commit.